+++ src/app/selfhelpInterfaces.ts
@@ -412,6 +412,7 @@ export interface MessageBoardStyle extends Style {
 
 export interface EntryListStyle extends Style {
     formName: StyleField,
+    load_as_table: StyleField,
 }
 
 export interface CalendarStyle extends Style {