+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - update `survey-angular-ui` from `1.9.109` to `1.9.125`
  - add an option to upload files to the server when uploaded with `survey-js`
  - add `save as pdf` for `survey-js`
+ - handle `edit_url` in `card` and show a button on the right side as in the web version
 
 ### Bugfix
  - #2 - fix the validation, `url` is correctly recognized with the capacitor.